requirements.txt
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.29
langchain_core==0.3.75
langchain_groq==0.3.7
langchain_text_splitters==0.3.9
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.46.1
sentence-transformers==5.1.0
